+++ semgrep/semgrep/cli.py
@@ -8,9 +8,11 @@ import semgrep.config_resolver
 import semgrep.semgrep_main
 import semgrep.test
 from semgrep import __VERSION__
+from semgrep.bytesize import parse_size
 from semgrep.constants import DEFAULT_CONFIG_FILE
 from semgrep.constants import DEFAULT_MAX_CHARS_PER_LINE
 from semgrep.constants import DEFAULT_MAX_LINES_PER_FINDING
+from semgrep.constants import DEFAULT_MAX_TARGET_SIZE
 from semgrep.constants import DEFAULT_TIMEOUT
 from semgrep.constants import MAX_CHARS_FLAG_NAME
 from semgrep.constants import MAX_LINES_FLAG_NAME
@@ -187,6 +189,18 @@ def cli() -> None:
         default=0,
         help=(
             "Maximum memory to use running a rule on a single file in MB. If set to 0 will not have memory limit. Defaults to 0."
+        ),
+    )
+
+    config.add_argument(
+        "--max-target-bytes",
+        type=parse_size,
+        default=DEFAULT_MAX_TARGET_SIZE,
+        help=(
+            "Maximum size for a file to be scanned by semgrep, e.g '1.5MB'. "
+            "Any input program larger than this will be ignored. "
+            "A zero or negative value disables this filter. "
+            f"Defaults to {DEFAULT_MAX_TARGET_SIZE} bytes."
         ),
     )
 
@@ -491,6 +505,7 @@ def cli() -> None:
                 jobs=args.jobs,
                 include=args.include,
                 exclude=args.exclude,
+                max_target_bytes=args.max_target_bytes,
                 strict=args.strict,
                 autofix=args.autofix,
                 dryrun=args.dryrun,